Add notification system with WebSocket support
- Introduced a new notifications table in the database to store user notifications.
- Implemented API handlers for listing notifications, marking them as read, and marking all as read.
- Created frontend components for displaying notifications and integrated them into the navigation.
- Enhanced UI to show notifications and allow users to manage their notification status.

+++ internal/db/migrations/000003_create_notifications_table.up.sql
@@ -2,7 +2,7 @@ CREATE TYPE notification_type AS ENUM ('info', 'success', 'warning', 'error');
 
 CREATE TABLE notifications (
     id BIGSERIAL PRIMARY KEY,
-    user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    user_id INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     title TEXT NOT NULL,
     message TEXT NOT NULL,
     type notification_type NOT NULL DEFAULT 'info',
