+++ internal/db/sqlc/models.go
@@ -86,7 +86,7 @@ type Mailbox struct {
 
 type Notification struct {
 	ID        int64            `json:"id"`
-	UserID    int64            `json:"user_id"`
+	UserID    int32            `json:"user_id"`
 	Title     string           `json:"title"`
 	Message   string           `json:"message"`
 	Type      NotificationType `json:"type"`
